pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bgm-nerdrock"
authors = [{ name = "Brian Gray", email = "bgraymusic@gmail.com" }]
version = "1.1"
requires-python = ">=3.12,<4.0"
readme = "README.md"
dependencies = [
    "deepmerge>=2.0,<3.0",
    "requests>=2.32.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "cryptography>=43.0.0,<44.0.0",
    "cffi>=1.17.0,<2.0.0",
]

[project.scripts]
bgnr-refresh = "api.runtime.database.database_handler:handle"

[project.optional-dependencies]
cdk = ["aws-cdk-lib>=2.171.1,<3.0.0", "constructs>=10.4.0,<11.0.0"]
test = [
    "boto3>=1.35.0,<2.0.0",
    "pytest>=8.3.0,<9.0.0",
    "pytest-mock>=3.14,<4.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "pytest-flake8>=1.3.0,<2.0.0",
    "pytest-pycodestyle>=2.4.0,<3.0.0",
    "pycodestyle>=2.12.0,<3.0.0",
    "flake8-pyproject>=1.2.0,<2.0.0",
]

[tool.setuptools]
include-package-data = true
packages = [
    "api",
    "api.runtime",
    "api.runtime.badges",
    "api.runtime.database",
    "api.runtime.discography",
]

[tool.setuptools.package-data]
"*" = ["*.yml"]

[tool.setuptools.exclude-package-data]
tests = ["*test*"]

[tool.pytest.ini_options]
minversion = "8.0"
required_plugins = ["pytest-cov", "pytest-flake8"]
addopts = "--import-mode=importlib --cov='api' --flake8"
pythonpath = "."
testpaths = ["api/test"]
python_files = "test_*.py"

[tool.coverage.run]
# include = ["api"]
omit = [
    "**/test/*",
    "**/infrastructure.py",
    "db/*",
    "web/*",
    "cdk.py",
    "refresh_database.py",
]
branch = true

[tool.coverage.report]
precision = 1
exclude_lines = ["^\\s*pass\\s*$", "^\\s*handle\\({}, {}\\)\\s*$"]
show_missing = true
fail_under = 80

[tool.flake8]
max-line-length = 120
exclude = ["api/test/**"]
